quick

+++ hw4/testing.c
@@ -10,7 +10,7 @@ int main() {
     printf("alloc is %p\n", front);
     printf("val is %d\n", *front);
     myfree(front);
-    int *nptr = mymalloc(sizeof(unsigned long));
+    unsigned long *nptr = mymalloc(sizeof(unsigned long));
     *nptr = 67;
     printf("for real it's %p %lu\n", nptr, *nptr);
     mycleanup();
